# doc/openapi.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Sentence aligner API"
  },
  "tags": [
    {
      "name": "sentences",
      "description": "Operations on sentences"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/sentences": {
      "post": {
        "tags": [
          "sentences"
        ],
        "summary": "Create a new sentence.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "The sentence object that has to be created.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Sentence"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The sentence has been created."
          },
          "409": {
            "description": "The given UUID is already used by another sentence."
          }
        }
      },
      "get": {
        "tags": [
          "sentences"
        ],
        "summary": "Get all the sentences.",
        "responses": {
          "200": {
            "description": "Returns all the sentences.",
            "schema": {
              "$ref": "#/definitions/sentences"
            },
            "examples": {
              "application/json": [
                {
                  "id": "4491-808e-405d-9fdd-0e853a383bcc",
                  "text": "It rains.",
                  "iso639_3": "eng",
                  "structure": "<sentence><subject>It</subject> <verb>rains</verb>.</sentence>"
                }
              ]
            }
          }
        }
      }
    },
    "/sentences/{sentence_id}": {
      "get": {
        "tags": [
          "sentences"
        ],
        "summary": "Returns one sentence properties.",
        "parameters": [
          {
            "name": "sentence_id",
            "in": "path",
            "description": "The sentence UUID.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Succeed. Returns the requested sentence content.",
            "schema": {
              "$ref": "#/definitions/Sentence"
            },
            "examples": {
              "application/json": {
                "id": "4491-808e-405d-9fdd-0e853a383bcc",
                "text": "It rains.",
                "iso639_3": "eng",
                "structure": "<sentence><subject>It</subject> <verb>rains</verb>.</sentence>"
              }
            }
          },
          "404": {
            "description": "The requested sentence cannot be found."
          }
        }
      }
    },
    "/sentences/{sentence_id}/text": {
      "put": {
        "tags": [
          "sentences"
        ],
        "summary": "Update the content of one sentence.",
        "parameters": [
          {
            "name": "sentence_id",
            "in": "path",
            "description": "The sentence UUID.",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "The new sentence content.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SentenceContent"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Succeed. The sentence content has been updated."
          },
          "404": {
            "description": "The requested sentence cannot be found."
          },
          "409": {
            "description": "A sentence with the same content and the same language already exists."
          }
        }
      }
    },
    "/sentences/{sentence_id}/language": {
      "put": {
        "tags": [
          "sentences"
        ],
        "summary": "Update the language of one sentence.",
        "parameters": [
          {
            "name": "sentence_id",
            "in": "path",
            "description": "The sentence UUID.",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "The new sentence language.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SentenceLanguage"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Succeed. The sentence language has been updated."
          },
          "404": {
            "description": "The requested sentence or the language cannot be found."
          },
          "409": {
            "description": "A sentence with identical language and content than the edited sentence already exists."
          }
        }
      }
    }
  },
  "definitions": {
    "Sentence": {
      "type": "object",
      "required": [
        "text",
        "iso639_3"
      ],
      "properties": {
        "id": {
          "type": "string",
          "example": "313a4491-808e-405d-9fdd-0e853a383bcc"
        },
        "text": {
          "$ref": "#/definitions/SentenceContent"
        },
        "iso639_3": {
          "$ref": "#/definitions/SentenceLanguage"
        },
        "structure": {
          "type": "string",
          "example": "<sentence><subject>It</subject> <verb>rains</verb>.</sentence>"
        }
      }
    },
    "SentenceContent": {
      "type": "string",
      "example": "It rains."
    },
    "SentenceLanguage": {
      "type": "string",
      "example": "eng"
    },
    "sentences": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Sentence"
      }
    }
  }
}